Refactor: Refactored plugin's commands design.
