+++ src/main/java/hugog/blockstreet/commands/ICommandFactory.java
@@ -1,0 +1,7 @@
+package hugog.blockstreet.commands;
+
+public interface ICommandFactory {
+    IPluginCommand produce(ICommandRegister commandRegister);
+    IPluginCommand produceMain();
+    IPluginCommand produceNull();
+}
