+++ src/main/java/hugog/blockstreet/commands/ICommandRegister.java
@@ -1,0 +1,6 @@
+package hugog.blockstreet.commands;
+
+public interface ICommandRegister {
+    String getAlias();
+    Class<? extends PluginCommand> getCommandExecutor();
+}
