+++ src/main/java/hugog/blockstreet/commands/IPluginCommand.java
@@ -1,0 +1,5 @@
+package hugog.blockstreet.commands;
+
+public interface IPluginCommand {
+    void execute();
+}
