+++ src/main/java/hugog/blockstreet/commands/implementation/DeleteCompanyCommand.java
@@ -1,6 +1,7 @@
-package hugog.blockstreet.commands;
+package hugog.blockstreet.commands.implementation;
 
 import hugog.blockstreet.Main;
+import hugog.blockstreet.commands.PluginCommand;
 import hugog.blockstreet.others.Company;
 import hugog.blockstreet.others.ConfigAccessor;
 import hugog.blockstreet.others.Messages;
@@ -10,55 +11,62 @@ import org.bukkit.entity.Player;
 
 import java.text.MessageFormat;
 
-public class DeleteCompanyCommand {
+/**
+ * Delete Company Command
+ *
+ * Command to delete companies in game.
+ *
+ * Syntax: /invest delete [ID]
+ *
+ * @author Hugo1307
+ * @version 1.0.0
+ */
+public class DeleteCompanyCommand extends PluginCommand {
 
-	private final String[] args;
-	private final CommandSender sender;
-	
-	public DeleteCompanyCommand(String[] args, CommandSender sender) {
-		this.args = args;
-		this.sender = sender;
+	public DeleteCompanyCommand(CommandSender sender, String[] args) {
+		super(sender, args);
 	}
-	
-	void runDeleteCompanyCmd() {
-		
+
+	@Override
+	public void execute() {
+
 		Player p = (Player) sender;
-		Messages messages = new Messages(Main.getInstance().messagesConfig);
-		
+		Messages messages = new Messages();
+
 		if (p.hasPermission("blockstreet.admin.delete") || p.hasPermission("blockstreet.admin.*")) {
-			
+
 			if (args.length > 1) {
-				
+
 				int companyId;
-				
+
 				try {
 					companyId = Integer.parseInt(args[1]);
 				} catch (NumberFormatException e) {
 					p.sendMessage(messages.getPluginPrefix() + messages.getWrongArguments());
 					return;
 				}
-				
+
 				Company companyToDelete = new Company(companyId).load();
-				
+
 				if (companyToDelete.exists()) {
-					
+
 					companyToDelete.delete(true);
-					
-					p.sendMessage(messages.getPluginPrefix() + MessageFormat.format(messages.getDeletedCompany(), 
+
+					p.sendMessage(messages.getPluginPrefix() + MessageFormat.format(messages.getDeletedCompany(),
 							ChatColor.GOLD + companyToDelete.getName() + ChatColor.GRAY));
-					
+
 				}else {
 					p.sendMessage(messages.getPluginPrefix() + messages.getInvalidCompany());
 				}
-			
+
 			}else {
 				p.sendMessage(messages.getPluginPrefix() + messages.getMissingArguments());
 			}
-			
+
 		}else {
 			p.sendMessage(messages.getPluginPrefix() + messages.getNoPermission());
 		}
-		
+
 	}
-	
+
 }
