+++ src/main/java/hugog/blockstreet/commands/implementation/InterestTimeCommand.java
@@ -1,6 +1,7 @@
-package hugog.blockstreet.commands;
+package hugog.blockstreet.commands.implementation;
 
 import hugog.blockstreet.Main;
+import hugog.blockstreet.commands.PluginCommand;
 import hugog.blockstreet.runnables.InterestRateRunnable;
 import hugog.blockstreet.others.Messages;
 import org.bukkit.command.CommandSender;
@@ -8,31 +9,40 @@ import org.bukkit.entity.Player;
 
 import java.text.MessageFormat;
 
-public class TimeLeftCommand {
+/**
+ * Interest Time Command
+ *
+ * Command to check how much time is left till the interests are paid.
+ *
+ * Syntax: /invest time
+ *
+ * @author Hugo1307
+ * @since v1.0.0
+ */
+public class InterestTimeCommand extends PluginCommand {
 
-	private final CommandSender sender;
-	
-	public TimeLeftCommand(CommandSender sender) {
-		this.sender = sender;
+	public InterestTimeCommand(CommandSender sender, String[] args) {
+		super(sender, args);
 	}
-	
-	void runTimeLeftCommand() {
-		
+
+	@Override
+	public void execute() {
+
 		Player p = (Player) sender;
-		Messages messages = new Messages(Main.getInstance().messagesConfig);
-		
+		Messages messages = new Messages();
+
 		if (p.hasPermission("blockstreet.command.time") || p.hasPermission("blockstreet.command.*")) {
-			
+
 			int totalMinutes = Main.getInstance().getConfig().getInt("BlockStreet.Timer");
 			int timeElapsed = InterestRateRunnable.getMinutesCounter();
 			int timeLeft = totalMinutes - timeElapsed;
-			
+
 			p.sendMessage(messages.getPluginPrefix() + MessageFormat.format(messages.getInterestRateTimeLeft(), timeLeft));
-			
-		}else {			
-			p.sendMessage(messages.getPluginPrefix() + messages.getNoPermission());			
+
+		}else {
+			p.sendMessage(messages.getPluginPrefix() + messages.getNoPermission());
 		}
-		
+
 	}
-	
+
 }
